backend/requirements.txt
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nibabel==5.2.1
numpy==1.26.4
packaging==24.0
python-dotenv==1.0.1
redis==5.0.4
scipy==1.12.0
Werkzeug==3.0.1
zipp==3.17.0
